requirements.txt
nose
nose-progressive
nose-testconfig
nose_parameterized
rednose
selenium
requests
flaky
pandas
xlrd
loguru
pyperclip
pymysql
git+https://github.com/0xIslamTaha/agent-python-nosetests@0.0.7
